Split runtime IK driver
